requirements.txt
pbr>=0.5.21,<1.0

Babel>=1.3
netaddr>=0.7.6
falcon>=0.1.6,<0.1.7
jsonschema>=1.3.0,!=1.4.0
iso8601>=0.1.8
msgpack-python
pymongo>=2.4
cassandra-driver>=1.0.0
python-keystoneclient>=0.4.1
python-memcached
WebOb>=1.2.3,<1.3
stevedore>=0.10
six>=1.4.1
oslo.config>=1.2.0
